requirements.txt
telethon>=1.30.0
python-telegram-bot>=20.0
cryptg>=0.4.0
python-socks[asyncio]>=2.0.0 